names in scratch-blockspecs need to be safe most importantly they can't have . [dot]
To actually test I added a simple comm-spec that connects to itself and thus effectively creates an echo-loopback. Surely beats the Bert-Ernie setup - and could be useful at all times for fast testing. (Why didn't we think of that one before?)

The attached new s2e file in this commit is one that is generated through the comm.js code now (and actually tested with the new echo-scratch-sample)

And I've removed the me.json (default launcher config) so everybody could have their own.